Rename JPetSmearingFunctionsNew to JPetSmearingFunctions
